# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * University Course Performance Tracker & Grade Prediction API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Student {
  id: number;
  name: string;
  email: string;
  student_id: string;
  year: number;
  major: string;
  /** @nullable */
  gpa?: number | null;
  created_at: string;
}

export interface StudentInput {
  name: string;
  email: string;
  student_id: string;
  year: number;
  major: string;
}

export interface StudentUpdate {
  name?: string;
  email?: string;
  year?: number;
  major?: string;
}

export interface CourseGradePoint {
  course_name: string;
  grade: string;
  grade_points: number;
  credits: number;
}

export interface StudentGpa {
  student_id: number;
  gpa: number;
  total_courses: number;
  completed_courses: number;
  letter_grade: string;
  grade_points_breakdown?: CourseGradePoint[];
}

export interface StudentCourseDetail {
  course_id: number;
  course_name: string;
  course_code: string;
  semester: string;
  /** @nullable */
  current_grade: number | null;
  /** @nullable */
  letter_grade: string | null;
  credits: number;
}

export interface Course {
  id: number;
  code: string;
  name: string;
  credits: number;
  semester: string;
  instructor: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  student_count?: number | null;
  /** @nullable */
  average_grade?: number | null;
  created_at: string;
}

export interface CourseInput {
  code: string;
  name: string;
  credits: number;
  semester: string;
  instructor: string;
  description?: string;
}

export interface CourseUpdate {
  name?: string;
  credits?: number;
  semester?: string;
  instructor?: string;
  description?: string;
}

export interface CourseStats {
  course_id: number;
  average_grade: number;
  median_grade: number;
  pass_rate: number;
  fail_rate: number;
  highest_grade: number;
  lowest_grade: number;
  student_count: number;
  /** @nullable */
  std_deviation?: number | null;
}

export interface Enrollment {
  id: number;
  student_id: number;
  course_id: number;
  semester: string;
  enrolled_at: string;
  /** @nullable */
  student_name?: string | null;
  /** @nullable */
  course_name?: string | null;
}

export interface EnrollmentInput {
  student_id: number;
  course_id: number;
  semester: string;
}

export interface Assignment {
  id: number;
  course_id: number;
  name: string;
  /** midterm|final|assignment|quiz|project */
  type: string;
  max_score: number;
  /** Weight 0-1 as percentage */
  weight: number;
  /** @nullable */
  due_date?: string | null;
  /** @nullable */
  description?: string | null;
  created_at: string;
}

export interface AssignmentInput {
  course_id: number;
  name: string;
  type: string;
  max_score: number;
  weight: number;
  due_date?: string;
  description?: string;
}

export interface AssignmentUpdate {
  name?: string;
  type?: string;
  max_score?: number;
  weight?: number;
  due_date?: string;
  description?: string;
}

export interface Grade {
  id: number;
  student_id: number;
  assignment_id: number;
  score: number;
  /** @nullable */
  percentage?: number | null;
  /** @nullable */
  letter_grade?: string | null;
  /** @nullable */
  feedback?: string | null;
  submitted_at: string;
  /** @nullable */
  student_name?: string | null;
  /** @nullable */
  assignment_name?: string | null;
  /** @nullable */
  course_id?: number | null;
}

export interface GradeInput {
  student_id: number;
  assignment_id: number;
  score: number;
  feedback?: string;
}

export interface GradeUpdate {
  score?: number;
  feedback?: string;
}

export interface GradeBucket {
  range: string;
  count: number;
  percentage?: number;
}

export interface LetterCount {
  letter: string;
  count: number;
  percentage?: number;
}

export interface GradeDistribution {
  course_id: number;
  buckets: GradeBucket[];
  letter_counts: LetterCount[];
}

export interface AtRiskStudent {
  student_id: number;
  student_name: string;
  current_grade: number;
  letter_grade: string;
  course_name: string;
  course_id?: number;
  assignments_missing?: number;
  /** high|medium|low */
  risk_level: string;
}

export interface CoursePerformance {
  course_id: number;
  course_name: string;
  average_grade: number;
  pass_rate: number;
  student_count: number;
  semester: string;
  /** @nullable */
  difficulty_score?: number | null;
}

export interface SemesterTrend {
  semester: string;
  gpa: number;
  courses_taken: number;
  /** @nullable */
  student_id?: number | null;
}

export interface AssignmentCompletion {
  assignment_id: number;
  assignment_name: string;
  type: string;
  submitted_count: number;
  total_enrolled: number;
  completion_rate: number;
  /** @nullable */
  average_score: number | null;
}

export interface GradePredictionInput {
  student_id: number;
  course_id: number;
  /** @nullable */
  midterm_score?: number | null;
  /** @nullable */
  assignment_completion_rate?: number | null;
}

export interface PredictionFactor {
  factor: string;
  weight: number;
  value: number;
}

export interface GradePrediction {
  student_id: number;
  course_id: number;
  predicted_score: number;
  predicted_letter: string;
  confidence: number;
  risk_level: string;
  factors?: PredictionFactor[];
}

export interface AtRiskPrediction {
  student_id: number;
  student_name: string;
  predicted_score: number;
  predicted_letter: string;
  risk_level: string;
  confidence: number;
  /** @nullable */
  current_score?: number | null;
}

export interface DashboardSummary {
  total_students: number;
  total_courses: number;
  total_enrollments: number;
  total_grades: number;
  average_gpa: number;
  at_risk_count: number;
  pass_rate: number;
  grade_distribution_overview?: LetterCount[];
}

export interface ActivityItem {
  id: number;
  /** grade_submitted|enrollment|assignment_created */
  type: string;
  description: string;
  timestamp: string;
  /** @nullable */
  entity_id?: number | null;
  /** @nullable */
  entity_type?: string | null;
}

export interface TopPerformer {
  student_id: number;
  student_name: string;
  gpa: number;
  major: string;
  year: number;
  rank?: number;
}

export type ListStudentsParams = {
search?: string;
major?: string;
year?: number;
};

export type ListCoursesParams = {
semester?: string;
search?: string;
};

export type ListEnrollmentsParams = {
student_id?: number;
course_id?: number;
semester?: string;
};

export type ListAssignmentsParams = {
course_id?: number;
type?: string;
};

export type ListGradesParams = {
student_id?: number;
course_id?: number;
assignment_id?: number;
};

export type GetGradeDistributionParams = {
course_id: number;
};

export type GetAtRiskStudentsParams = {
course_id?: number;
/**
 * GPA threshold, defaults to 2.0
 */
threshold?: number;
};

export type GetSemesterTrendsParams = {
student_id?: number;
};

export type GetAssignmentCompletionParams = {
course_id?: number;
};

export type GetRecentActivityParams = {
limit?: number;
};

export type GetTopPerformersParams = {
limit?: number;
};