# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: University Course Performance Tracker & Grade Prediction API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - in: query
          name: semester
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /courses/{id}/students:
    get:
      operationId: getCourseStudents
      tags: [courses]
      summary: List students enrolled in a course
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Enrolled students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
  /courses/{id}/stats:
    get:
      operationId: getCourseStats
      tags: [courses]
      summary: Get grade stats for a course
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Course statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseStats"

  # ─── Enrollments ─────────────────────────────────────────────────
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent grade submissions and enrollments
      parameters:
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Recent activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/top-performers:
    get:
      operationId: getTopPerformers
      tags: [dashboard]
      summary: Get top-performing students
      parameters:
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Top performers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopPerformer"
components:
  schemas:
...
    Student:
      type: object
      required: [id, name, email, student_id, year, major, created_at]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        student_id: { type: string }
        year: { type: integer }
        major: { type: string }
        gpa: { type: ["number", "null"] }
        created_at: { type: string }
...
    Course:
      type: object
      required: [id, code, name, credits, semester, instructor, created_at]
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        credits: { type: integer }
        semester: { type: string }
        instructor: { type: string }
        description: { type: ["string", "null"] }
        student_count: { type: ["integer", "null"] }
        average_grade: { type: ["number", "null"] }
        created_at: { type: string }
    CourseInput:
      type: object
      required: [code, name, credits, semester, instructor]
      properties:
        code: { type: string }
        name: { type: string }
        credits: { type: integer }
        semester: { type: string }
        instructor: { type: string }
        description: { type: string }
    CourseUpdate:
      type: object
      properties:
        name: { type: string }
        credits: { type: integer }
        semester: { type: string }
        instructor: { type: string }
        description: { type: string }
    CourseStats:
      type: object
      required: [course_id, average_grade, median_grade, pass_rate, fail_rate, highest_grade, lowest_grade, student_count]
      properties:
        course_id: { type: integer }
        average_grade: { type: number }
        median_grade: { type: number }
        pass_rate: { type: number }
        fail_rate: { type: number }
        highest_grade: { type: number }
        lowest_grade: { type: number }
        student_count: { type: integer }
        std_deviation: { type: ["number", "null"] }

    # ── Enrollment ────────────────────────────────────────────────────
...
    ActivityItem:
      type: object
      required: [id, type, description, timestamp]
      properties:
        id: { type: integer }
        type: { type: string, description: "grade_submitted|enrollment|assignment_created" }
        description: { type: string }
        timestamp: { type: string }
        entity_id: { type: ["integer", "null"] }
        entity_type: { type: ["string", "null"] }
    TopPerformer:
      type: object
      required: [student_id, student_name, gpa, major, year]
      properties:
        student_id: { type: integer }
        student_name: { type: string }
        gpa: { type: number }
        major: { type: string }
        year: { type: integer }
        rank: { type: integer }