# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * University Course Performance Tracker & Grade Prediction API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  /** grade_submitted|enrollment|assignment_created */
  type: string;
  description: string;
  timestamp: string;
  /** @nullable */
  entity_id?: number | null;
  /** @nullable */
  entity_type?: string | null;
}