# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * University Course Performance Tracker & Grade Prediction API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  email: string;
  student_id: string;
  year: number;
  major: string;
  /** @nullable */
  gpa?: number | null;
  created_at: string;
}